requirements.txt
biopython==1.85
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.4
PyMuPDF==1.25.5
requests==2.32.3
toml==0.10.2
urllib3==2.3.0
